# Inplace Encryption for Blob Granule