Uses in-place encryption for blob granule. See below for performance results from the `bench_serialize_deltas` microbenchmark (tests the encryption codepath mainly):

Without inplace encryption:
```
bench_serialize_deltas/131072/32768/1      162302 ns       162303 ns         4297 bytes_per_second=769M/s serialized_bytes=409M 
bench_serialize_deltas/524288/32768/1      933581 ns      933587 ns          750 bytes_per_second=535M/s serialized_bytes=307M 
bench_serialize_deltas/1048576/32768/1    2218665 ns     2218673 ns          316 bytes_per_second=450M/s serialized_bytes=268M
```

With inplace encryption:
```
bench_serialize_deltas/131072/32768/1      161153 ns       161153 ns         4339 bytes_per_second=775M/s serialized_bytes=412M - **around 1% improvement**
bench_serialize_deltas/524288/32768/1      917204 ns     917210 ns          764 bytes_per_second=544M/s serialized_bytes=312M  - **around 2% improvement**
bench_serialize_deltas/1048576/32768/1   2190447 ns    2190452 ns          319 bytes_per_second=456M/s serialized_bytes=271M - **around 2% improvement**
```

Overall we notice around a 2% improvement with inplace encryption enabled (each run above represents a different granule delta size and a fixed chunk size pf 32KB)

# Code-Reviewer Section

The general pull request guidelines can be found [here](https://github.com/apple/foundationdb/wiki/FoundationDB-Commit-Process).

Please check each of the following things and check *all* boxes before accepting a PR.

- [ ] The PR has a description, explaining both the problem and the solution.
- [ ] The description mentions which forms of testing were done and the testing seems reasonable.
- [ ] Every function/class/actor that was touched is reasonably well documented.

## For Release-Branches

If this PR is made against a release-branch, please also check the following:

- [ ] This change/bugfix is a cherry-pick from the next younger branch (younger `release-branch` or `main` if this is the youngest branch)
- [ ] There is a good reason why this PR needs to go into a release branch and this reason is documented (either in the description above or in a linked GitHub issue)